backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
SQLAlchemy==2.0.40
alembic==1.15.2
PyMySQL==1.1.1
redis==5.2.1
pydantic-settings==2.8.1
PyJWT==2.10.1
bcrypt==4.0.1
